Added Spreadsheet demonstrating concept
Using spreadsheets can be a great way to visualize performance. When you compare the number of cells occupied on the two separate sheets you can get a feeling for how many calculations are required. Ask, "How many additional cells will be required if we add another line of data, if we double the amount of data?"